Feature-View Design pattern change using route grouping
