+++ src/app/(view)/Frames/Footer/EmailSignUpInput.tsx
