+++ src/app/(view)/Frames/Footer/index.tsx
