+++ src/app/(view)/components/AiCard/AiChip.tsx
