+++ src/app/(view)/components/AiCard/index.tsx
