+++ src/app/(view)/components/Image/index.tsx
