+++ src/app/(view)/components/SearchInput/index.tsx
