+++ src/app/(view)/page.tsx
@@ -1,0 +1,15 @@
+import DiscoverLatestAi from "./Frames/DiscoverLatestAi";
+import HeroFrame from "./Frames/HeroFrame";
+import MasterYourAi from "./Frames/MasterYourAi";
+
+export default function Home() {
+  return (
+    <div>
+      <HeroFrame />
+
+      <DiscoverLatestAi />
+
+      <MasterYourAi />
+    </div>
+  );
+}
