--- src/app/page.tsx
@@ -1,21 +1,0 @@
-import DiscoverLatestAi from "./Frames/DiscoverLatestAi";
-import Footer from "./Frames/Footer";
-import Header from "./Frames/Header";
-import HeroFrame from "./Frames/HeroFrame";
-import MasterYourAi from "./Frames/MasterYourAi";
-
-export default function Home() {
-    return (
-        <div>
-            <Header />
-
-            <HeroFrame />
-
-            <DiscoverLatestAi />
-
-            <MasterYourAi />
-
-            <Footer />
-        </div>
-    );
-}
